requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
black==22.1.0
click==8.0.4
Django==4.0.2
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
